Add static assertions for size and alignment to generated Rust code.
Followup CLs will add corresponding assertions to the generated C++ code as well
as assertions for field offsets.

PiperOrigin-RevId: 398421352

+++ rs_bindings_from_cc/ir.rs
@@ -66,8 +66,8 @@ pub struct Field {
 pub struct Record {
     pub identifier: Identifier,
     pub fields: Vec<Field>,
-    pub size: i64,
-    pub alignment: i64,
+    pub size: usize,
+    pub alignment: usize,
 }
 
 #[derive(Debug, Default, PartialEq, Eq, Hash, Clone, Deserialize)]
